pyproject.toml
[project]
name = "mybooks"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django>=5.2.1",
  "django-admin-interface>=0.30.0",
  "psutil>=6.1.1",
  "whitenoise>=6.9.0",
  "gunicorn>=23.0.0",
  "django-components>=0.139.1",
  "django-extensions>=4.1",
  "rust-just>=1.4.0",
  "pyopenssl>=25.1.0",
  "python-dotenv>=1.1.0",
  "pydantic>=2.11.7",
  "raven>=6.10.0",
  "sentry-sdk[django]>=2.38.0",
  "logfire[django,sqlite3]>=4.9.0",
  "django-tailwind[reload]>=4.2.0",
  "django-health-check>=3.20.0",
  "django-oauth-toolkit>=3.0.1",
  "djangorestframework>=3.16.1",
  "django-oauth-toolkit-dcr>=0.1.0",
  "django-cors-headers>=4.9.0",
  "drf-spectacular>=0.28.0",
]

[dependency-groups]
dev = [
  "Werkzeug[watchdog]",
  "ipdb",
  "flake8",
  "flake8-isort",
  "black",
  "pylint-django",
  "doppler-env",
  "django-debug-toolbar",
  "pre-commit",
  "ipython",
  "djlint",
  "debugpy",
  "pylint-django",
  "prospector",
  "autoflake",
  "ruff",
  "ty",
  "django-types",
  "honcho>=2.0.0",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 150
target-version = ['py313']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 150
known_first_party = ["app"]
skip = [".venv/"]
skip_glob = ["**/migrations/*.py"]

[tool.pylint.MASTER]
ignore = '.git,.venv,node_modules,*/migrations/*'
load-plugins = ["pylint_django"]
django-settings-module = "mybooks.settings"

[tool.pylint.FORMAT]
max-line-length = 150

[tool.pylint."MESSAGES CONTROL"]
disable = ["missing-docstring", "invalid-name", "import-outside-toplevel", "too-few-public-methods", "unspecified-encoding", "too-many-locals", "redefined-builtin", "line-too-long"]

[tool.pylint.DESIGN]
max-parents = 13

[tool.pylint.TYPECHECK]
generated-members = [
  "REQUEST",
  "acl_users",
  "aq_parent",
  "[a-zA-Z]+_set{1,2}",
  "save",
  "delete",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".github",
  ".hg",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "build",
  "dist",
  "migrations",
  "node_modules",
  "static",
]
# Same as Black.
line-length = 150
# Assume Python 3.12.
target-version = "py313"

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
ignore = ["E501", "E741"]                # temporary
per-file-ignores = {}
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
unfixable = []

[tool.djlint]
# H017: Void tags should be self closing.
# H031:	Consider adding meta keywords.
ignore = "H017,H031"
preserve_blank_lines = true
indent = 2

# ==== autoflake ====
[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
ignore-init-module-imports = true
expand-star-imports = true
exclude = ["*/migrations/*", ".venv", "node_modules"]
